requirements.txt
# Smart Canary Protocol - Python Dependencies

# Core functionality
requests>=2.25.1
python-dotenv>=0.19.0
pyyaml>=6.0

# AI and machine learning
openai>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data sources and analysis
feedparser>=6.0.0
yfinance>=0.1.70
beautifulsoup4>=4.10.0
markdown2>=2.4.0

# Visualization and reporting
matplotlib>=3.5.0

# Economic data APIs
fredapi>=0.4.3

# Date and time handling
python-dateutil>=2.8.0

# Optional integrations
slack-sdk>=3.15.0  # For Slack notifications
praw>=7.5.0        # For Reddit API (optional)

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
